Split message event
