+++ ccd-definition/CaseEventToFields/CareSupervision/messageJudge.json
@@ -3,43 +3,6 @@
     "LiveFrom": "01/01/2017",
     "CaseTypeID": "CARE_SUPERVISION_EPO",
     "CaseEventID": "messageJudgeOrLegalAdviser",
-    "CaseFieldID": "hasJudicialMessages",
-    "PageFieldDisplayOrder": 1,
-    "DisplayContext": "READONLY",
-    "PageID": "messageJudgeOptions",
-    "ShowSummaryChangeOption": "N",
-    "PageDisplayOrder": 1,
-    "FieldShowCondition": "messageJudgeOption=\"DO_NOT_SHOW\"",
-    "CallBackURLMidEvent": "${CCD_DEF_CASE_SERVICE_BASE_URL}/callback/message-judge/mid-event"
-  },
-  {
-    "LiveFrom": "01/01/2017",
-    "CaseTypeID": "CARE_SUPERVISION_EPO",
-    "CaseEventID": "messageJudgeOrLegalAdviser",
-    "CaseFieldID": "messageJudgeOption",
-    "PageFieldDisplayOrder": 2,
-    "DisplayContext": "MANDATORY",
-    "PageID": "messageJudgeOptions",
-    "ShowSummaryChangeOption": "N",
-    "PageDisplayOrder": 1,
-    "PageShowCondition": "hasJudicialMessages=\"Yes\""
-  },
-  {
-    "LiveFrom": "01/01/2017",
-    "CaseTypeID": "CARE_SUPERVISION_EPO",
-    "CaseEventID": "messageJudgeOrLegalAdviser",
-    "CaseFieldID": "judicialMessageDynamicList",
-    "PageFieldDisplayOrder": 3,
-    "DisplayContext": "MANDATORY",
-    "PageID": "messageJudgeOptions",
-    "ShowSummaryChangeOption": "N",
-    "PageDisplayOrder": 1,
-    "FieldShowCondition": "messageJudgeOption=\"REPLY\""
-  },
-  {
-    "LiveFrom": "01/01/2017",
-    "CaseTypeID": "CARE_SUPERVISION_EPO",
-    "CaseEventID": "messageJudgeOrLegalAdviser",
     "CaseFieldID": "hasAdditionalApplications",
     "PageFieldDisplayOrder": 1,
     "DisplayContext": "READONLY",
@@ -94,8 +57,7 @@
     "DisplayContext": "READONLY",
     "PageID": "NewMessageNote",
     "ShowSummaryChangeOption": "N",
-    "PageDisplayOrder": 3,
-    "CallBackURLMidEvent": "${CCD_DEF_CASE_SERVICE_BASE_URL}/callback/message-judge/validate/mid-event"
+    "PageDisplayOrder": 3
   },
   {
     "LiveFrom": "01/01/2017",
@@ -108,18 +70,6 @@
     "ShowSummaryChangeOption": "N",
     "PageDisplayOrder": 3,
     "FieldShowCondition": "isMessageRegardingAdditionalApplications=\"Yes\""
-  },
-  {
-    "LiveFrom": "01/01/2017",
-    "CaseTypeID": "CARE_SUPERVISION_EPO",
-    "CaseEventID": "messageJudgeOrLegalAdviser",
-    "CaseFieldID": "judicialMessageReply",
-    "PageFieldDisplayOrder": 3,
-    "DisplayContext": "COMPLEX",
-    "PageID": "NewMessageNote",
-    "ShowSummaryChangeOption": "Y",
-    "PageDisplayOrder": 3,
-    "FieldShowCondition": "messageJudgeOption=\"REPLY\""
   },
   {
     "LiveFrom": "01/01/2017",
