+++ service/src/test/java/uk/gov/hmcts/reform/fpl/service/ReplyToMessageJudgeServiceTest.java
@@ -69,7 +69,7 @@ import static uk.gov.hmcts.reform.fpl.utils.TestDataHelper.testDocumentReference
 
 @ExtendWith({MockitoExtension.class})
 @MockitoSettings(strictness = Strictness.LENIENT)
-class MessageJudgeServiceTest {
+class ReplyToMessageJudgeServiceTest {
 
     private static final String COURT_EMAIL = "ctsc@test.com";
     private static final String MESSAGE_NOTE = "Message note";
@@ -86,15 +86,13 @@ class MessageJudgeServiceTest {
     @Mock
     private Time time;
     @Mock
-    private IdentityService identityService;
-    @Mock
     private UserService userService;
     @Mock
     private CtscEmailLookupConfiguration ctscEmailLookupConfiguration;
     @Spy
     private ObjectMapper mapper = new ObjectMapper();
     @InjectMocks
-    private MessageJudgeService messageJudgeService;
+    private ReplyToMessageJudgeService replyToMessageJudgeService;
 
     @BeforeEach
     void init() {
@@ -148,7 +146,7 @@ class MessageJudgeServiceTest {
                 .build())
             .build();
 
-        final Map<String, Object> expectedEventData = messageJudgeService.initialiseCaseFields(caseData);
+        final Map<String, Object> expectedEventData = replyToMessageJudgeService.initialiseCaseFields(caseData);
 
         final DynamicList expectedAdditionalApplicationsDynamicList = buildDynamicList(
             Pair.of(c2DocumentBundle.getId(), "C2, 01 Dec 2020"),
@@ -192,7 +190,7 @@ class MessageJudgeServiceTest {
             .additionalApplicationsBundle(additionalApplicationsBundle)
             .build();
 
-        Map<String, Object> data = messageJudgeService.initialiseCaseFields(caseData);
+        Map<String, Object> data = replyToMessageJudgeService.initialiseCaseFields(caseData);
 
         DynamicList expectedAdditionalApplicationsDynamicList = buildDynamicList(
             Pair.of(applicationId, "C2, 01 Dec 2020")
@@ -228,7 +226,7 @@ class MessageJudgeServiceTest {
             .judicialMessages(judicialMessages)
             .build();
 
-        Map<String, Object> data = messageJudgeService.initialiseCaseFields(caseData);
+        Map<String, Object> data = replyToMessageJudgeService.initialiseCaseFields(caseData);
 
         DynamicList expectedJudicialDynamicList = buildDynamicList(
             Pair.of(judicialMessages.get(0).getId(), "01 Dec 2020"),
@@ -269,7 +267,7 @@ class MessageJudgeServiceTest {
             .closedJudicialMessages(closedJudicialMessages)
             .build();
 
-        Map<String, Object> data = messageJudgeService.initialiseCaseFields(caseData);
+        Map<String, Object> data = replyToMessageJudgeService.initialiseCaseFields(caseData);
 
         DynamicList expectedJudicialDynamicList = buildDynamicList(
             Pair.of(judicialMessages.get(0).getId(), "01 Dec 2020")
@@ -288,7 +286,7 @@ class MessageJudgeServiceTest {
 
     @Test
     void shouldPopulateOnlyEmailAddressesWhenDocumentsDoNotExist() {
-        assertThat(messageJudgeService.initialiseCaseFields(CaseData.builder().build()))
+        assertThat(replyToMessageJudgeService.initialiseCaseFields(CaseData.builder().build()))
             .containsExactly(
                 entry("judicialMessageMetaData", JudicialMessageMetaData.builder()
                     .sender(COURT_EMAIL)
@@ -302,7 +300,7 @@ class MessageJudgeServiceTest {
 
         CaseData caseData = CaseData.builder().build();
 
-        assertThat(messageJudgeService.initialiseCaseFields(caseData))
+        assertThat(replyToMessageJudgeService.initialiseCaseFields(caseData))
             .containsExactly(
                 entry("judicialMessageMetaData", JudicialMessageMetaData.builder()
                     .recipient(COURT_EMAIL)
@@ -315,201 +313,11 @@ class MessageJudgeServiceTest {
 
         CaseData caseData = CaseData.builder().build();
 
-        assertThat(messageJudgeService.initialiseCaseFields(caseData))
+        assertThat(replyToMessageJudgeService.initialiseCaseFields(caseData))
             .containsExactly(
                 entry("judicialMessageMetaData", JudicialMessageMetaData.builder()
                     .sender(COURT_EMAIL)
                     .recipient(EMPTY).build()));
-    }
-
-    @Test
-    void shouldRebuildAdditionalApplicationDynamicListAndFormatDocumentsCorrectlyWhenOtherApplicationSelected() {
-        UUID c2DocumentBundleId = randomUUID();
-
-        OtherApplicationsBundle selectedOtherDocumentBundle = OtherApplicationsBundle.builder()
-            .id(SELECTED_DYNAMIC_LIST_ITEM_ID)
-            .uploadedDateTime("1 January 2021, 12:00pm")
-            .applicationType(OtherApplicationType.C1_APPOINTMENT_OF_A_GUARDIAN)
-            .document(DocumentReference.builder()
-                .filename(OTHER_FILE_NAME)
-                .build())
-            .supportingEvidenceBundle(List.of(
-                element(SupportingEvidenceBundle.builder()
-                    .document(DocumentReference.builder()
-                        .filename(OTHER_SUPPORTING_DOCUMENT_FILE_NAME)
-                        .build())
-                    .build())))
-            .build();
-
-        C2DocumentBundle c2DocumentBundle = C2DocumentBundle.builder()
-            .id(c2DocumentBundleId)
-            .uploadedDateTime("2 January 2021, 12:00pm")
-            .document(DocumentReference.builder()
-                .filename(C2_FILE_NAME)
-                .build())
-            .build();
-
-        AdditionalApplicationsBundle additionalApplicationsBundle = AdditionalApplicationsBundle.builder()
-            .otherApplicationsBundle(selectedOtherDocumentBundle)
-            .c2DocumentBundle(c2DocumentBundle)
-            .build();
-
-        MessageJudgeEventData messageJudgeEventData = MessageJudgeEventData.builder()
-            .additionalApplicationsDynamicList(DynamicList.builder()
-                .value(DynamicListElement.builder()
-                    .code(SELECTED_DYNAMIC_LIST_ITEM_ID)
-                    .build())
-                .build())
-            .build();
-
-        CaseData caseData = CaseData.builder()
-            .messageJudgeEventData(messageJudgeEventData)
-            .additionalApplicationsBundle(wrapElements(additionalApplicationsBundle))
-            .build();
-
-        String expectedRelatedDocumentsLabel = OTHER_FILE_NAME + "\n" + OTHER_SUPPORTING_DOCUMENT_FILE_NAME;
-
-        assertThat(messageJudgeService.populateNewMessageFields(caseData))
-            .extracting("relatedDocumentsLabel", "additionalApplicationsDynamicList")
-            .containsExactly(
-                expectedRelatedDocumentsLabel,
-                buildDynamicList(0,
-                    Pair.of(SELECTED_DYNAMIC_LIST_ITEM_ID, "C1, 1 January 2021, 12:00pm"),
-                    Pair.of(c2DocumentBundleId, "C2, 2 January 2021, 12:00pm")
-                )
-            );
-    }
-
-    @Test
-    void shouldRebuildAdditionalApplicationDynamicListAndFormatDocumentsCorrectlyWhenC2ApplicationSelected() {
-        UUID otherDocumentBundleId = randomUUID();
-
-        OtherApplicationsBundle selectedOtherDocumentBundle = OtherApplicationsBundle.builder()
-            .id(otherDocumentBundleId)
-            .uploadedDateTime("1 January 2021, 12:00pm")
-            .applicationType(OtherApplicationType.C1_APPOINTMENT_OF_A_GUARDIAN)
-            .document(DocumentReference.builder()
-                .filename(OTHER_FILE_NAME)
-                .build())
-            .build();
-
-        C2DocumentBundle c2DocumentBundle = C2DocumentBundle.builder()
-            .id(SELECTED_DYNAMIC_LIST_ITEM_ID)
-            .uploadedDateTime("2 January 2021, 12:00pm")
-            .document(DocumentReference.builder()
-                .filename(C2_FILE_NAME)
-                .build())
-            .supportingEvidenceBundle(List.of(
-                element(SupportingEvidenceBundle.builder()
-                    .document(DocumentReference.builder()
-                        .filename(C2_SUPPORTING_DOCUMENT_FILE_NAME)
-                        .build())
-                    .build())))
-            .build();
-
-        AdditionalApplicationsBundle additionalApplicationsBundle = AdditionalApplicationsBundle.builder()
-            .otherApplicationsBundle(selectedOtherDocumentBundle)
-            .c2DocumentBundle(c2DocumentBundle)
-            .build();
-
-        MessageJudgeEventData messageJudgeEventData = MessageJudgeEventData.builder()
-            .additionalApplicationsDynamicList(DynamicList.builder()
-                .value(DynamicListElement.builder()
-                    .code(SELECTED_DYNAMIC_LIST_ITEM_ID)
-                    .build())
-                .build())
-            .build();
-
-        CaseData caseData = CaseData.builder()
-            .messageJudgeEventData(messageJudgeEventData)
-            .additionalApplicationsBundle(wrapElements(additionalApplicationsBundle))
-            .build();
-
-        String expectedRelatedDocumentsLabel = C2_FILE_NAME + "\n" + C2_SUPPORTING_DOCUMENT_FILE_NAME;
-
-        assertThat(messageJudgeService.populateNewMessageFields(caseData))
-            .extracting("relatedDocumentsLabel", "additionalApplicationsDynamicList")
-            .containsExactly(
-                expectedRelatedDocumentsLabel,
-                buildDynamicList(1,
-                    Pair.of(otherDocumentBundleId, "C1, 1 January 2021, 12:00pm"),
-                    Pair.of(SELECTED_DYNAMIC_LIST_ITEM_ID, "C2, 2 January 2021, 12:00pm")
-                )
-            );
-    }
-
-    @Test
-    void shouldReturnEmptyMapWhenC2DocumentHasNotBeenSelected() {
-        C2DocumentBundle selectedC2DocumentBundle = C2DocumentBundle.builder()
-            .document(DocumentReference.builder()
-                .filename(C2_FILE_NAME)
-                .build())
-            .supportingEvidenceBundle(List.of(
-                element(SupportingEvidenceBundle.builder()
-                    .document(DocumentReference.builder()
-                        .filename(C2_SUPPORTING_DOCUMENT_FILE_NAME)
-                        .build())
-                    .build())))
-            .build();
-
-        CaseData caseData = CaseData.builder()
-            .c2DocumentBundle(List.of(
-                element(SELECTED_DYNAMIC_LIST_ITEM_ID, selectedC2DocumentBundle),
-                element(randomUUID(), C2DocumentBundle.builder()
-                    .document(DocumentReference.builder()
-                        .filename("other_c2.doc")
-                        .build())
-                    .build())
-            ))
-            .build();
-
-        assertThat(messageJudgeService.populateNewMessageFields(caseData)).isEmpty();
-    }
-
-    @Test
-    void shouldReturnEmptyMapWhenAdditionalApplicationDocumentHasNotBeenSelected() {
-        C2DocumentBundle selectedC2DocumentBundle = C2DocumentBundle.builder()
-            .document(DocumentReference.builder()
-                .filename(C2_FILE_NAME)
-                .build())
-            .supportingEvidenceBundle(List.of(
-                element(SupportingEvidenceBundle.builder()
-                    .document(DocumentReference.builder()
-                        .filename(C2_SUPPORTING_DOCUMENT_FILE_NAME)
-                        .build())
-                    .build())))
-            .build();
-
-        C2DocumentBundle notSelectedC2DocumentBundle = C2DocumentBundle.builder()
-            .id(randomUUID())
-            .document(DocumentReference.builder()
-                .filename("other_c2.doc")
-                .build())
-            .build();
-
-        CaseData caseData = CaseData.builder()
-            .additionalApplicationsBundle(
-                wrapElements(
-                    AdditionalApplicationsBundle.builder()
-                        .c2DocumentBundle(selectedC2DocumentBundle)
-                        .build(),
-                    AdditionalApplicationsBundle.builder()
-                        .c2DocumentBundle(notSelectedC2DocumentBundle)
-                        .build()
-                )
-            )
-            .build();
-
-        assertThat(messageJudgeService.populateNewMessageFields(caseData)).isEmpty();
-    }
-
-    @Test
-    void shouldNotPrePopulateRecipientWhenMessageIsInitiatedNotByJudge() {
-        when(userService.hasUserRole(UserRole.JUDICIARY)).thenReturn(false);
-
-        CaseData caseData = CaseData.builder().build();
-
-        assertThat(messageJudgeService.populateNewMessageFields(caseData)).isEmpty();
     }
 
     @Test
@@ -554,7 +362,7 @@ class MessageJudgeServiceTest {
             .latestMessage("")
             .build();
 
-        Map<String, Object> populateReplyMessageFields = messageJudgeService.populateReplyMessageFields(caseData);
+        Map<String, Object> populateReplyMessageFields = replyToMessageJudgeService.populateReplyMessageFields(caseData);
 
         assertThat(populateReplyMessageFields)
             .extracting("judicialMessageReply", "judicialMessageDynamicList")
@@ -610,7 +418,7 @@ class MessageJudgeServiceTest {
             .latestMessage("")
             .build();
 
-        Map<String, Object> populateReplyMessageFields = messageJudgeService.populateReplyMessageFields(caseData);
+        Map<String, Object> populateReplyMessageFields = replyToMessageJudgeService.populateReplyMessageFields(caseData);
 
         assertThat(populateReplyMessageFields)
             .extracting("judicialMessageReply")
@@ -637,312 +445,9 @@ class MessageJudgeServiceTest {
                 element(JudicialMessage.builder().build())))
             .build();
 
-        assertThatThrownBy(() -> messageJudgeService.populateReplyMessageFields(caseData))
+        assertThatThrownBy(() -> replyToMessageJudgeService.populateReplyMessageFields(caseData))
             .isInstanceOf(JudicialMessageNotFoundException.class)
             .hasMessage(format("Judicial message with id %s not found", SELECTED_DYNAMIC_LIST_ITEM_ID));
-    }
-
-    @Test
-    void shouldAppendNewJudicialMessageToJudicialMessageListWhenDocumentNotSelected() {
-        JudicialMessageMetaData judicialMessageMetaData = JudicialMessageMetaData.builder()
-            .subject(MESSAGE_REQUESTED_BY)
-            .sender(MESSAGE_SENDER)
-            .recipient(MESSAGE_RECIPIENT)
-            .urgency("High urgency")
-            .build();
-
-        MessageJudgeEventData messageJudgeEventData = MessageJudgeEventData.builder()
-            .judicialMessageNote(MESSAGE_NOTE)
-            .judicialMessageMetaData(judicialMessageMetaData)
-            .build();
-
-        CaseData caseData = CaseData.builder().messageJudgeEventData(messageJudgeEventData).build();
-
-        when(identityService.generateId()).thenReturn(NEW_ELEMENT_ID);
-        when(userService.getUserEmail()).thenReturn(MESSAGE_SENDER);
-
-        List<Element<JudicialMessage>> updatedMessages = messageJudgeService.addNewJudicialMessage(caseData);
-
-        Element<JudicialMessage> expectedJudicialMessageElement = element(NEW_ELEMENT_ID, JudicialMessage.builder()
-            .updatedTime(time.now())
-            .status(OPEN)
-            .latestMessage(MESSAGE_NOTE)
-            .dateSent(formatLocalDateTimeBaseUsingFormat(time.now(), DATE_TIME_AT))
-            .sender(MESSAGE_SENDER)
-            .recipient(MESSAGE_RECIPIENT)
-            .subject(MESSAGE_REQUESTED_BY)
-            .urgency("High urgency")
-            .messageHistory(String.format("%s - %s", MESSAGE_SENDER, MESSAGE_NOTE))
-            .build());
-
-        assertThat(updatedMessages).hasSize(1).first().isEqualTo(expectedJudicialMessageElement);
-    }
-
-    @Test
-    void shouldAppendNewJudicialMessageToJudicialMessageListWhenAdditionalApplicationDocumentHasBeenSelected() {
-        JudicialMessageMetaData judicialMessageMetaData = JudicialMessageMetaData.builder()
-            .recipient(MESSAGE_RECIPIENT)
-            .build();
-
-        DocumentReference mainC2DocumentReference = DocumentReference.builder()
-            .filename(C2_FILE_NAME)
-            .build();
-
-        DocumentReference supportingC2DocumentReference = DocumentReference.builder()
-            .filename(C2_SUPPORTING_DOCUMENT_FILE_NAME)
-            .build();
-
-        C2DocumentBundle selectedC2DocumentBundle = C2DocumentBundle.builder()
-            .id(SELECTED_DYNAMIC_LIST_ITEM_ID)
-            .document(mainC2DocumentReference)
-            .supportingEvidenceBundle(List.of(
-                element(SupportingEvidenceBundle.builder()
-                    .document(supportingC2DocumentReference)
-                    .build())))
-            .build();
-
-        MessageJudgeEventData messageJudgeEventData = MessageJudgeEventData.builder()
-            .judicialMessageNote(MESSAGE_NOTE)
-            .judicialMessageMetaData(judicialMessageMetaData)
-            .additionalApplicationsDynamicList(DynamicList.builder()
-                .value(DynamicListElement.builder()
-                    .code(SELECTED_DYNAMIC_LIST_ITEM_ID)
-                    .build())
-                .build())
-            .build();
-
-        CaseData caseData = CaseData.builder()
-            .messageJudgeEventData(messageJudgeEventData)
-            .additionalApplicationsBundle(wrapElements(AdditionalApplicationsBundle.builder()
-                .c2DocumentBundle(selectedC2DocumentBundle)
-                .build()))
-            .build();
-
-        when(userService.getUserEmail()).thenReturn(MESSAGE_SENDER);
-
-        List<Element<JudicialMessage>> updatedMessages = messageJudgeService.addNewJudicialMessage(caseData);
-        JudicialMessage newMessage = updatedMessages.get(0).getValue();
-        List<Element<DocumentReference>> relatedDocuments = newMessage.getRelatedDocuments();
-
-        assertThat(newMessage.getRelatedDocumentFileNames()).isEqualTo(
-            selectedC2DocumentBundle.getAllC2DocumentFileNames()
-        );
-        assertThat(relatedDocuments.get(0).getValue()).isEqualTo(mainC2DocumentReference);
-        assertThat(relatedDocuments.get(1).getValue()).isEqualTo(supportingC2DocumentReference);
-    }
-
-    @Test
-    void shouldAppendNewJudicialMessageToJudicialMessageListWhenOtherApplicationDocumentHasBeenSelected() {
-        JudicialMessageMetaData judicialMessageMetaData = JudicialMessageMetaData.builder()
-            .recipient(MESSAGE_RECIPIENT)
-            .build();
-
-        DocumentReference mainC2DocumentReference = DocumentReference.builder()
-            .filename(C2_FILE_NAME)
-            .build();
-
-        DocumentReference supportingC2DocumentReference = DocumentReference.builder()
-            .filename(C2_SUPPORTING_DOCUMENT_FILE_NAME)
-            .build();
-
-        OtherApplicationsBundle selectedOtherApplicationBundle = OtherApplicationsBundle.builder()
-            .id(SELECTED_DYNAMIC_LIST_ITEM_ID)
-            .applicationType(OtherApplicationType.C1_APPOINTMENT_OF_A_GUARDIAN)
-            .document(mainC2DocumentReference)
-            .supportingEvidenceBundle(List.of(
-                element(SupportingEvidenceBundle.builder()
-                    .document(supportingC2DocumentReference)
-                    .build())))
-            .build();
-
-        MessageJudgeEventData messageJudgeEventData = MessageJudgeEventData.builder()
-            .judicialMessageNote(MESSAGE_NOTE)
-            .judicialMessageMetaData(judicialMessageMetaData)
-            .additionalApplicationsDynamicList(DynamicList.builder()
-                .value(DynamicListElement.builder()
-                    .code(SELECTED_DYNAMIC_LIST_ITEM_ID)
-                    .build())
-                .build())
-            .build();
-
-        CaseData caseData = CaseData.builder()
-            .messageJudgeEventData(messageJudgeEventData)
-            .additionalApplicationsBundle(wrapElements(AdditionalApplicationsBundle.builder()
-                .otherApplicationsBundle(selectedOtherApplicationBundle)
-                .build()))
-            .build();
-
-        when(userService.getUserEmail()).thenReturn(MESSAGE_SENDER);
-
-        List<Element<JudicialMessage>> updatedMessages = messageJudgeService.addNewJudicialMessage(caseData);
-        JudicialMessage newMessage = updatedMessages.get(0).getValue();
-        List<Element<DocumentReference>> relatedDocuments = newMessage.getRelatedDocuments();
-
-        assertThat(newMessage.getRelatedDocumentFileNames()).isEqualTo(
-            selectedOtherApplicationBundle.getAllDocumentFileNames()
-        );
-        assertThat(relatedDocuments.get(0).getValue()).isEqualTo(mainC2DocumentReference);
-        assertThat(relatedDocuments.get(1).getValue()).isEqualTo(supportingC2DocumentReference);
-    }
-
-    @Test
-    void shouldAppendNewJudicialMessageToJudicialMessageListWhenPlacementApplicationHasBeenSelected() {
-
-        final DocumentReference placementApplication = testDocumentReference("placement application");
-        final DocumentReference supportingDocument1 = testDocumentReference("first supporting document");
-        final DocumentReference supportingDocument2 = testDocumentReference("second supporting document");
-        final DocumentReference confidentialDocument1 = testDocumentReference("first confidential document");
-        final DocumentReference notice1 = testDocumentReference("first notice document");
-        final DocumentReference notice1Response = testDocumentReference("first notice response");
-
-        final Element<Placement> placement = element(Placement.builder()
-            .childName("Alex Green")
-            .placementUploadDateTime(LocalDateTime.of(2020, 10, 12, 13, 0))
-            .application(placementApplication)
-            .placementNotice(notice1)
-            .supportingDocuments(wrapElements(
-                PlacementSupportingDocument.builder()
-                    .document(supportingDocument1)
-                    .build(),
-                PlacementSupportingDocument.builder()
-                    .document(supportingDocument2)
-                    .build()))
-            .confidentialDocuments(wrapElements(
-                PlacementConfidentialDocument.builder()
-                    .document(confidentialDocument1)
-                    .build()))
-            .noticeDocuments(wrapElements(
-                PlacementNoticeDocument.builder()
-                    .response(notice1Response)
-                    .build()))
-            .build());
-
-        final MessageJudgeEventData messageJudgeEventData = MessageJudgeEventData.builder()
-            .judicialMessageNote(MESSAGE_NOTE)
-            .judicialMessageMetaData(JudicialMessageMetaData.builder()
-                .recipient(MESSAGE_RECIPIENT)
-                .build())
-            .additionalApplicationsDynamicList(DynamicList.builder()
-                .value(DynamicListElement.builder()
-                    .code(placement.getId())
-                    .build())
-                .build())
-            .build();
-
-        final CaseData caseData = CaseData.builder()
-            .messageJudgeEventData(messageJudgeEventData)
-            .placementEventData(PlacementEventData.builder()
-                .placements(List.of(placement))
-                .build())
-            .build();
-
-        final List<Element<JudicialMessage>> updatedMessages = messageJudgeService.addNewJudicialMessage(caseData);
-
-        final JudicialMessage newMessage = updatedMessages.get(0).getValue();
-
-        assertThat(newMessage.getRelatedDocumentFileNames()).isEqualTo(
-            "Application: placement application\n"
-                + "Supporting documents: first supporting document, second supporting document\n"
-                + "Confidential documents: first confidential document\n"
-                + "Notice: first notice document\n"
-                + "Notice responses: first notice response");
-
-        assertThat(newMessage.getRelatedDocuments()).extracting(Element::getValue).containsExactly(
-            placementApplication,
-            supportingDocument1,
-            supportingDocument2,
-            confidentialDocument1,
-            notice1,
-            notice1Response);
-    }
-
-    @Test
-    void shouldAppendJudicialMessageToJudicialMessageListWhenPlacementApplicationWithMissingDocumentsHasBeenSelected() {
-
-        final DocumentReference placementApplication = testDocumentReference("placement application");
-
-        final Element<Placement> placement = element(Placement.builder()
-            .childName("Alex Green")
-            .placementUploadDateTime(LocalDateTime.of(2020, 10, 12, 13, 0))
-            .application(placementApplication)
-            .build());
-
-        final MessageJudgeEventData messageJudgeEventData = MessageJudgeEventData.builder()
-            .judicialMessageNote(MESSAGE_NOTE)
-            .judicialMessageMetaData(JudicialMessageMetaData.builder()
-                .recipient(MESSAGE_RECIPIENT)
-                .build())
-            .additionalApplicationsDynamicList(DynamicList.builder()
-                .value(DynamicListElement.builder()
-                    .code(placement.getId())
-                    .build())
-                .build())
-            .build();
-
-        final CaseData caseData = CaseData.builder()
-            .messageJudgeEventData(messageJudgeEventData)
-            .placementEventData(PlacementEventData.builder()
-                .placements(List.of(placement))
-                .build())
-            .build();
-
-        final List<Element<JudicialMessage>> updatedMessages = messageJudgeService.addNewJudicialMessage(caseData);
-
-        final JudicialMessage newMessage = updatedMessages.get(0).getValue();
-
-        assertThat(newMessage.getRelatedDocumentFileNames())
-            .isEqualTo("Application: placement application");
-
-        assertThat(newMessage.getRelatedDocuments())
-            .extracting(Element::getValue)
-            .containsExactly(placementApplication);
-    }
-
-    @Test
-    void shouldAppendNewJudicialMessageToExistingJudicialMessageList() {
-        JudicialMessage newMessage = JudicialMessage.builder()
-            .subject(MESSAGE_REQUESTED_BY)
-            .recipient(MESSAGE_RECIPIENT)
-            .sender(MESSAGE_SENDER)
-            .build();
-
-        UUID existingJudicialMessageId = randomUUID();
-        JudicialMessage existingJudicialMessage = JudicialMessage.builder().build();
-
-        List<Element<JudicialMessage>> existingJudicialMessages = new ArrayList<>();
-        existingJudicialMessages.add(element(existingJudicialMessageId, existingJudicialMessage));
-
-        MessageJudgeEventData messageJudgeEventData = MessageJudgeEventData.builder()
-            .judicialMessageNote(MESSAGE_NOTE)
-            .judicialMessageMetaData(newMessage)
-            .build();
-
-        CaseData caseData = CaseData.builder()
-            .messageJudgeEventData(messageJudgeEventData)
-            .judicialMessages(existingJudicialMessages)
-            .build();
-
-        JudicialMessage expectedNewJudicialMessage = JudicialMessage.builder()
-            .sender(MESSAGE_SENDER)
-            .recipient(MESSAGE_RECIPIENT)
-            .updatedTime(time.now())
-            .status(OPEN)
-            .subject(MESSAGE_REQUESTED_BY)
-            .latestMessage(MESSAGE_NOTE)
-            .messageHistory(String.format("%s - %s", MESSAGE_SENDER, MESSAGE_NOTE))
-            .dateSent(formatLocalDateTimeBaseUsingFormat(time.now(), DATE_TIME_AT))
-            .build();
-
-        when(identityService.generateId()).thenReturn(NEW_ELEMENT_ID);
-        when(userService.getUserEmail()).thenReturn(MESSAGE_SENDER);
-
-        List<Element<JudicialMessage>> updatedMessages = messageJudgeService.addNewJudicialMessage(caseData);
-
-        assertThat(updatedMessages.size()).isEqualTo(2);
-        assertThat(updatedMessages).isEqualTo(List.of(
-            element(existingJudicialMessageId, existingJudicialMessage),
-            element(NEW_ELEMENT_ID, expectedNewJudicialMessage)
-        ));
     }
 
     @Test
@@ -964,7 +469,7 @@ class MessageJudgeServiceTest {
             .judicialMessages(List.of(element(SELECTED_DYNAMIC_LIST_ITEM_ID, buildJudicialMessage(dateSent))))
             .build();
 
-        assertThat(messageJudgeService.validateJudgeReplyMessage(caseData))
+        assertThat(replyToMessageJudgeService.validateJudgeReplyMessage(caseData))
             .containsExactly("The sender's and recipient's email address cannot be the same");
     }
 
@@ -987,7 +492,7 @@ class MessageJudgeServiceTest {
             .judicialMessages(List.of(element(SELECTED_DYNAMIC_LIST_ITEM_ID, buildJudicialMessage(dateSent))))
             .build();
 
-        assertThat(messageJudgeService.validateJudgeReplyMessage(caseData)).isEmpty();
+        assertThat(replyToMessageJudgeService.validateJudgeReplyMessage(caseData)).isEmpty();
     }
 
     @ParameterizedTest
@@ -1010,7 +515,7 @@ class MessageJudgeServiceTest {
             .judicialMessages(List.of(element(SELECTED_DYNAMIC_LIST_ITEM_ID, buildJudicialMessage(dateSent))))
             .build();
 
-        assertThat(messageJudgeService.validateJudgeReplyMessage(caseData)).isEmpty();
+        assertThat(replyToMessageJudgeService.validateJudgeReplyMessage(caseData)).isEmpty();
     }
 
     @ParameterizedTest
@@ -1031,7 +536,7 @@ class MessageJudgeServiceTest {
             .judicialMessages(List.of(element(SELECTED_DYNAMIC_LIST_ITEM_ID, buildJudicialMessage(dateSent))))
             .build();
 
-        assertThat(messageJudgeService.validateJudgeReplyMessage(caseData)).isEmpty();
+        assertThat(replyToMessageJudgeService.validateJudgeReplyMessage(caseData)).isEmpty();
     }
 
     @Test
@@ -1050,7 +555,7 @@ class MessageJudgeServiceTest {
 
         when(userService.getUserEmail()).thenReturn(MESSAGE_RECIPIENT);
 
-        Map<String, Object> updatedData = messageJudgeService.updateJudicialMessages(caseData);
+        Map<String, Object> updatedData = replyToMessageJudgeService.updateJudicialMessages(caseData);
         List<Element<JudicialMessage>> updatedMessages =
             (List<Element<JudicialMessage>>) updatedData.get("judicialMessages");
 
@@ -1087,7 +592,7 @@ class MessageJudgeServiceTest {
         judicialMessages.add(pastJudicialMessage);
 
         List<Element<JudicialMessage>> sortedJudicialMessages
-            = messageJudgeService.sortJudicialMessages(judicialMessages);
+            = replyToMessageJudgeService.sortJudicialMessages(judicialMessages);
 
         assertThat(sortedJudicialMessages).isEqualTo(List.of(latestJudicialMessage, pastJudicialMessage,
             oldestJudicialMessage));
@@ -1105,7 +610,7 @@ class MessageJudgeServiceTest {
                 .build())))
             .build();
 
-        assertThat(messageJudgeService.getNextHearingLabel(caseData))
+        assertThat(replyToMessageJudgeService.getNextHearingLabel(caseData))
             .isEqualTo(String.format("Next hearing in the case: %s hearing, %s", hearingType.getLabel(),
                 formatLocalDateTimeBaseUsingFormat(hearingStartDate, DATE)));
     }
@@ -1114,7 +619,7 @@ class MessageJudgeServiceTest {
     void shouldNotPopulateFirstHearingLabelWhenHearingDoesNotExists() {
         CaseData caseData = CaseData.builder().build();
 
-        assertThat(messageJudgeService.getNextHearingLabel(caseData)).isEmpty();
+        assertThat(replyToMessageJudgeService.getNextHearingLabel(caseData)).isEmpty();
     }
 
     @Test
@@ -1139,7 +644,7 @@ class MessageJudgeServiceTest {
         when(userService.getUserEmail()).thenReturn(MESSAGE_RECIPIENT);
         List<Element<JudicialMessage>> expectedJudicialMessages = List.of(oldJudicialMessage);
 
-        Map<String, Object> updatedJudicialMessages = messageJudgeService.updateJudicialMessages(caseData);
+        Map<String, Object> updatedJudicialMessages = replyToMessageJudgeService.updateJudicialMessages(caseData);
 
         assertThat(updatedJudicialMessages).containsEntry("judicialMessages", expectedJudicialMessages);
 
@@ -1166,7 +671,7 @@ class MessageJudgeServiceTest {
                 element(randomUUID(), JudicialMessage.builder().build())))
             .build();
 
-        assertThatThrownBy(() -> messageJudgeService.updateJudicialMessages(caseData))
+        assertThatThrownBy(() -> replyToMessageJudgeService.updateJudicialMessages(caseData))
             .isInstanceOf(JudicialMessageNotFoundException.class)
             .hasMessage(format("Judicial message with id %s not found", SELECTED_DYNAMIC_LIST_ITEM_ID));
     }
